Update Dockerfile for clearlinux
Add a workaround for docker bug https://github.com/moby/moby/issues/31805

When using the overlay storage driver, a file modification is not detected by Docker when the file's modification time remains unchanged. 
Once dockerhub switches to graph driver that does not have this limitation, this workaround is not needed.